# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/products/{slug}": {
      "get": {
        "operationId": "ProductsController_getBySlug",
        "summary": "Get a single ACTIVE product by slug.",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "products (public)"
        ]
      }
    },
...
    "/api/admin/products": {
      "get": {
        "operationId": "AdminProductsController_list",
        "summary": "List all products (any status) for the admin table.",
        "parameters": [
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "Field to sort by (allow-listed per resource).",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "schema": {
              "default": "desc",
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Full-text-ish search over name/description.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category",
            "required": false,
            "in": "query",
            "description": "Category slug filter.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "minPrice",
            "required": false,
            "in": "query",
            "description": "Minimum price in cents.",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "maxPrice",
            "required": false,
            "in": "query",
            "description": "Maximum price in cents.",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "DRAFT",
                "ACTIVE",
                "ARCHIVED"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "products (admin)"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "AdminProductsController_create",
        "summary": "Create a product with variants + images.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "products (admin)"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/admin/products/{id}": {
      "get": {
        "operationId": "AdminProductsController_get",
        "summary": "Get a product by id (admin).",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "products (admin)"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "operationId": "AdminProductsController_update",
        "summary": "Update a product (reconciles variants + categories).",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "products (admin)"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "AdminProductsController_remove",
        "summary": "Delete a product.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "products (admin)"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/admin/products/bulk/status": {
      "patch": {
        "operationId": "AdminProductsController_bulkStatus",
        "summary": "Bulk change product status.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "products (admin)"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/admin/products/{id}/images": {
      "post": {
        "operationId": "AdminProductsController_addImage",
        "summary": "Attach an uploaded image URL to a product.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "products (admin)"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/admin/products/{id}/images/reorder": {
      "patch": {
        "operationId": "AdminProductsController_reorderImages",
        "summary": "Reorder product images.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReorderImagesDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "products (admin)"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/admin/products/{id}/images/{imageId}": {
      "delete": {
        "operationId": "AdminProductsController_deleteImage",
        "summary": "Delete a product image.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "imageId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "products (admin)"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/checkout": {
      "post": {
        "operationId": "OrdersController_checkout",
        "summary": "Create an order from the current cart (server recomputes totals + stock).",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckoutDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "checkout & orders"
        ]
      }
    },
    "/api/orders/{id}/pay": {
      "post": {
        "operationId": "OrdersController_pay",
        "summary": "Stubbed payment — marks the order PAID. // TODO: real PSP.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "checkout & orders"
        ]
      }
    },
    "/api/orders/{orderNumber}": {
      "get": {
        "operationId": "OrdersController_getByNumber",
        "summary": "Look up an order by its number (confirmation page).",
        "parameters": [
          {
            "name": "orderNumber",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "checkout & orders"
        ]
      }
    },
    "/api/account/orders": {
      "get": {
        "operationId": "OrdersController_myOrders",
        "summary": "List the signed-in customer’s orders.",
        "parameters": [
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "Field to sort by (allow-listed per resource).",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "schema": {
              "default": "desc",
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "PENDING",
                "PAID",
                "FULFILLED",
                "CANCELLED",
                "REFUNDED"
              ],
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Search order number or email.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "checkout & orders"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/account/orders/{orderNumber}": {
      "get": {
        "operationId": "OrdersController_myOrder",
        "summary": "Get one of the signed-in customer’s orders.",
        "parameters": [
          {
            "name": "orderNumber",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "checkout & orders"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/admin/orders": {
      "get": {
        "operationId": "AdminOrdersController_list",
        "summary": "List/filter orders (status, search) with pagination.",
        "parameters": [
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "Field to sort by (allow-listed per resource).",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "schema": {
              "default": "desc",
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "PENDING",
                "PAID",
                "FULFILLED",
                "CANCELLED",
                "REFUNDED"
              ],
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Search order number or email.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "orders (admin)"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/admin/orders/{id}": {
      "get": {
        "operationId": "AdminOrdersController_get",
        "summary": "Order detail with line items + totals + address.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "orders (admin)"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/admin/orders/{id}/status": {
      "patch": {
        "operationId": "AdminOrdersController_updateStatus",
        "summary": "Update order status (PENDING/PAID/FULFILLED/CANCELLED/REFUNDED).",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrderStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "orders (admin)"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Nexlor Commerce API",
    "description": "Storefront + Admin API for the Nexlor commerce template.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "VariantInputDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "priceOverride": {
            "type": "object",
            "description": "Override price in cents; null = use basePrice."
          }
        }
      },
      "ImageInputDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "alt": {
            "type": "string"
          }
        }
      },
      "CreateProductDto": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string",
            "description": "URL slug; auto-generated from name if omitted."
          },
          "basePrice": {
            "type": "number",
            "description": "Base price in cents."
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "ACTIVE",
              "ARCHIVED"
            ],
            "default": "DRAFT"
          },
          "categoryIds": {
            "description": "Category ids.",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "variants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VariantInputDto"
            }
          },
          "images": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ImageInputDto"
            }
          }
        }
      },
      "UpdateProductDto": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string",
            "description": "URL slug; auto-generated from name if omitted."
          },
          "basePrice": {
            "type": "number",
            "description": "Base price in cents."
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "ACTIVE",
              "ARCHIVED"
            ],
            "default": "DRAFT"
          },
          "categoryIds": {
            "description": "Category ids.",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "variants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VariantInputDto"
            }
          },
          "images": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ImageInputDto"
            }
          }
        }
      },
      "BulkStatusDto": {
        "type": "object",
        "properties": {}
      },
      "ReorderImagesDto": {
        "type": "object",
        "properties": {}
      },
...
      "AddressDto": {
        "type": "object",
        "properties": {}
      },
      "CheckoutDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "buyer@example.com"
          },
          "billingAddress": {
            "$ref": "#/components/schemas/AddressDto"
          }
        },
        "required": [
          "email"
        ]
      },
      "UpdateOrderStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "PAID",
              "FULFILLED",
              "CANCELLED",
              "REFUNDED"
            ]
          }
        },
        "required": [
          "status"
        ]
      }
    }
  }
}